Making redo.erl's state self updating
The last stable version is v1.1.0, in its own branch. However, moving
from that branch to the current master in production cases may break
due to the internal tuple change, and that not all redo users have it
under supervision, the update should ideally be done internally.

This patch adds internal checks for the State argument and calls for an
update if it detects the old form from v1.1.0 and updates it to the
current form, setting the reconnection argument to 'true'.